requirements.txt
torch~=2.3.1
torch-directml~=0.2.3.dev240715
numpy~=1.26.4
h5py~=3.13.0
torchvision~=0.18.1
pillow~=10.4.0
tqdm~=4.66.5
scikit_learn~=1.5.1
scipy~=1.14.1
matplotlib~=3.9.2
requests~=2.32.3
click~=8.1.7
timm~=1.0.9
scikit-learn~=1.5.2
streamlit~=1.45.0